Honor -n to disable the database without exploding
Also makes rails honor a database.yml given on the commandline.
Hierarchy is now: commandline option, user's home config, installer
config, and finally current working directory + "/config/database.yml"
(the rails default).

Also moves all requires in msfconsole below option parsing so -h is fast
again.

After this commit, it is now possible to start msfconsole with -n again,
but **only in production mode** because of a dep on FactoryGirl in
development.

+++ config/application.rb
@@ -1,29 +1,36 @@
 require 'rails'
 require File.expand_path('../boot', __FILE__)
 
-# only the parts of 'rails/all' that metasploit-framework actually uses
-begin
-  require 'active_record/railtie'
-rescue LoadError
-  warn "activerecord not in the bundle, so database support will be disabled."
-  warn "Bundle installed '--without #{Bundler.settings.without.join(' ')}'"
-  warn "To clear the without option do `bundle install --without ''` " \
-       "(the --without flag with an empty string) or " \
-       "`rm -rf .bundle` to remove the .bundle/config manually and " \
-       "then `bundle install`"
+$msfconsole_options ||= {}
+
+# only attempt to load active_record if database support is enabled
+unless $msfconsole_options['DisableDatabase']
+  # only the parts of 'rails/all' that metasploit-framework actually uses
+  begin
+    require 'active_record/railtie'
+  rescue LoadError
+    warn "activerecord not in the bundle, so database support will be disabled."
+    warn "Bundle installed '--without #{Bundler.settings.without.join(' ')}'"
+    warn "To clear the without option do `bundle install --without ''` " \
+      "(the --without flag with an empty string) or " \
+      "`rm -rf .bundle` to remove the .bundle/config manually and " \
+      "then `bundle install`"
+  end
 end
 
 all_environments = [
-    :development,
-    :production,
-    :test
+  :development,
+  :production,
+  :test
 ]
 
+db_environments = ($msfconsole_options['DisableDatabase'] ? [] : all_environments)
+
 Bundler.require(
-    *Rails.groups(
-        db: all_environments,
-        pcap: all_environments
-    )
+  *Rails.groups(
+    db: db_environments,
+    pcap: all_environments
+  )
 )
 
 #
@@ -38,20 +45,25 @@ module Metasploit
     class Application < Rails::Application
       include Metasploit::Framework::CommonEngine
 
-      user_config_root = Pathname.new(Msf::Config.get_config_root)
-      user_database_yaml = user_config_root.join('database.yml')
-
-      # First try the user's configuration
-      if user_database_yaml.exist?
-        config.paths['config/database'] = [user_database_yaml.to_path]
+      if $msfconsole_options['DatabaseYAML'] && File.exists?($msfconsole_options['DatabaseYAML'])
+        config.paths['config/database'] = [ $msfconsole_options['DatabaseYAML'] ]
       else
-        # That didn't work out, try the config created by the installer.
-        install_root = Pathname.new(Msf::Config.install_root)
-        installer_database_yml = install_root.parent.join('ui').join('config').join('database.yml')
-        if installer_database_yml.exist?
-          config.paths['config/database'] = [installer_database_yaml.to_path]
+        user_config_root = Pathname.new(Msf::Config.get_config_root)
+        user_database_yaml = user_config_root.join('database.yml')
+
+        # First try the user's configuration
+        if user_database_yaml.exist?
+          config.paths['config/database'] = [ user_database_yaml.to_path ]
+        else
+          # That didn't work out, try the config created by the installer.
+          install_root = Pathname.new(Msf::Config.install_root)
+          installer_database_yml = install_root.parent.join('ui').join('config').join('database.yml')
+          if installer_database_yml.exist?
+            config.paths['config/database'] = [ installer_database_yaml.to_path ]
+          end
         end
       end
+
     end
   end
 end
